# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Pronia shop API - Customer
  description: A sample API to demonstrate Swagger documentation with schema relationships.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local server 
...
paths:
...
  /users:
    get:
      tags: 
        - Customer    
      summary: Get the current user's profile
      description: Retrieve the profile of the currently logged in user.
      responses:
        '200':
          description: A JSON object of the user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile not found"
    
    post:
      tags: 
        - Customer    
      summary: Register a new user
      description: Register a new user with the provided details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - email
                - password
              properties:
                fullName:
                  type: string
                  example: "Alice Jones"
                  description: The full name of the user.
                email:
                  type: string
                  format: email
                  example: "alice.jones@example.com"
                  description: The email address of the user.
                password:
                  type: string
                  minLength: 6
                  example: "password123"
                  description: The password for the user account. Must be at least 6 characters long.
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  userId:
                    type: string
                    example: "1234567890"
                    description: The unique identifier for the newly registered user.
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid user details"
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                    description: Details about validation errors.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An unexpected error occurred"        

    put:
      tags: 
        - Customer    
      summary: Update the user's profile
      description: Update the profile of the currently logged-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  example: "Alice Jones"     
                address:
                  type: string
                  example: "789 Oak St"
                phone:
                  type: string
                  example: "555-7890"
                dob: 
                  type: string
                  example: "1992-03-10"              
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid profile details"
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                    description: Details about validation errors.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized access"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An unexpected error occurred"
components:
  schemas:
...
    permission_group:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: "Customer"
        permissions:
          type: array
          # items:
          #   $ref: '#/components/schemas/Permission'
          example: [ 
                {"resource": "post", "action": { "create": false, "read": true, "update": false, "delete": false } }, 
                { "resource": "comment", "action": { "create": true, "read": true, "update": true, "delete": true }}]
        active:
          type: boolean
          example: true
    users:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 3
        fullName:
          type: string
          example: "Alice Jones"
        email: 
          type: string
          example: "alice.jones@example.com"
        # password:
        #   type: string
        #   example: "$2b$10$zVQ.FB77IECAfCt7uENASO4bAV8IKIhGQJJGMbIuNmYP70LzBkbCC"
        permission_group:
          $ref: '#/components/schemas/permission_group'
          # type: integer
          # format: int32
          # example: 1
        address:
          type: string
          example: "789 Oak St"
        phone:
          type: string
          example: "555-7890"
        dob: 
          type: string
          example: "1992-03-10"
        lastLogin:
          type: string
          example: "2021-01-01 12:00:00"
        createdAt:
          type: string
          example: "2021-01-01 12:00:00"
        updatedAt:
          type: string
          example: "2021-01-01 12:00:00"
        active:
          type: boolean
          example: true